use unsigned for timer times fixed return value of ProcessImpl::wait

+++ src/Pt-System/Timer.cpp
@@ -45,7 +45,7 @@ size_t Timer::interval() const
 }
 
 
-void Timer::setInterval(size_t msecs)
+void Timer::setInterval(unsigned msecs)
 {
     _interval = msecs;
